typecast i to uint16_t in GetMaximumProcessorCount(i);

+++ libs/platform/user/ebpf_platform_user.cpp
@@ -262,7 +262,7 @@ ebpf_platform_initiate()
         uint32_t base_index = 0;
         for (size_t i = 0; i < _ebpf_platform_group_to_index_map.size(); i++) {
             _ebpf_platform_group_to_index_map[i] = base_index;
-            base_index += GetMaximumProcessorCount(i);
+            base_index += GetMaximumProcessorCount((uint16_t)i);
         }
     } catch (...) {
         return EBPF_NO_MEMORY;
